kernelbase: Fix GetEnvironmentVariableA with multi-byte strings.
The previous implementation of GetEnvironmentVariableA incorrectly
assumed that, after WideCharToMultiByte conversion, the resulting A
string will have the same number of CHARs as the number of WCHARs in
the W string. This assumption resulted in failures when querying
environment variables with non-ASCII values in locales with multi-byte
character pages.

Address this by always retrieving the Unicode value of the variable,
then measuring its length after code page conversion, and only then
deciding whether it fits into the user-supplied buffer.

Since we are no longer preallocating a buffer with a user-specified
size, the 32767 character limit on the size parameter is removed, as
it should be enforced naturally by the size of the environment block.

Signed-off-by: Vladimir Panteleev <git@vladimir.panteleev.md>
---

Since v3: New patch to fix a deficiency observed during implementation
of previous commit.

I haven't found how to get the test bot to run these tests against
Wine. Selecting "Japanese:Japan" for the debian10 VM still causes the
test suite to run in an environment where GetACP() still returns 1252.

The tests do run locally if I run `LANG=ja_JP.UTF-8 make environ.ok`,
and pass.